Check if the marker size is valid
Set the pose_calculation property to false if the marker size is less than zero.
This avoid break the system when:
Have id-to-size list ->  Detect a non-listed id -> The marker_size property is set to -1 (default value).

For that case, the behavior now is detect the marker (draw its borders), but do not compute its pose.